requirements.txt
asgiref==3.2.10
beautifulsoup4==4.9.3
certifi==2020.6.20
cloudinary==1.22.0
confusable-homoglyphs==3.2.0
Django==3.1.2
django-bootstrap==0.2.4
django-bootstrap4==2.3.1
django-registration==3.1.1
djangorestframework==3.12.1
Pillow==8.0.1
psycopg2==2.8.6
pytz==2020.1
six==1.15.0
soupsieve==2.0.1
sqlparse==0.4.1
urllib3==1.25.11
whitenoise==5.2.0
